Separated encoded adapters
